parser/requirements.txt
et_xmlfile==2.0.0
lxml==6.0.2
openpyxl==3.1.5
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
python-docx==1.2.0
typing_extensions==4.15.0
